+++ packages/ub/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - new `Session.setTempTenantID(newTenantID)` method - allow switch tenant ID "on the fly" (require server 5.22.8)
+ - new `Session.runInTenant(tenantID, func)` method - allow execution code in context of another tenant
  - added Deutsch (`de`) locale translations
 
 ### Changed
@@ -16,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.22.35] - 2022-05-04
 ## [5.22.34] - 2022-04-29
 ## [5.22.33] - 2022-04-27
 ## [5.22.32] - 2022-04-25